Move classes to appropriate package.
